Refactor AllTimeBlock and CollectionsCardList Move to separate folder, improve readable and create custom hooks
