+++ src/pages/DashboardPage/components/DashboardLendTab/components/AllTimeBlock/constants.ts
@@ -3,17 +3,17 @@ export enum AllTimeStatus {
   Defaulted = 'totalDefaulted',
 }
 
-export const ALL_TIME_COLOR_MAP: Record<AllTimeStatus, string> = {
+export const STATUS_COLOR_MAP: Record<AllTimeStatus, string> = {
   [AllTimeStatus.Repaid]: 'var(--additional-green-primary)',
   [AllTimeStatus.Defaulted]: 'var(--additional-red-primary)',
 }
 
-export const ALL_TIME_DISPLAY_NAMES: Record<AllTimeStatus, string> = {
+export const STATUS_DISPLAY_NAMES: Record<AllTimeStatus, string> = {
   [AllTimeStatus.Repaid]: 'Total repaid',
   [AllTimeStatus.Defaulted]: 'Total defaulted',
 }
 
-export const EMPTY_SINGLE_BAR_CHART_DATA = {
+export const NO_DATA_CHART = {
   key: 'empty',
   label: 'Empty',
   value: 100,
