+++ src/pages/DashboardPage/components/DashboardLendTab/components/AllTimeBlock/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AllTimeBlock'
